+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 27. Updated AssetRipper.Primitives to 3.2.0
 28. Updated Il2CppInterop to 1.5.0-ci.643
 29. Fixed an issue with Melon Subfolder toggle being inverted
+30. Fixed an issue with Il2CppAssemblyGenerator and File Detection Failures
 
 ---